+++ mocket/mocketizer.py
@@ -1,4 +1,4 @@
-from mocket.mocket import Mocket
+from mocket.core.mocket import Mocket
 from mocket.mode import MocketMode
 from mocket.utils import get_mocketize
