+++ mocket/mode.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Any, ClassVar
 
 from mocket.core.exceptions import StrictMocketException
-from mocket.mocket import Mocket
+from mocket.core.mocket import Mocket
 
 if TYPE_CHECKING:  # pragma: no cover
     from typing import NoReturn
